% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.miBMA.R
\name{mice.impute.miBMA}
\alias{mice.impute.miBMA}
\title{mice.impute.miBMA Function}
\usage{
mice.impute.miBMA(y, ry, x, wy = NULL, ...)
}
\description{
This function allows you to run miBMA
}
\author{
this is a different version of mice.impute.norm which was written
by Stef van Buuren, Karin Groothuis-Oudshoorn, 2017
}
