% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.bicreg.R
\name{my.bicreg}
\alias{my.bicreg}
\title{Specialized bicreg function}
\usage{
my.bicreg(x, y, wt = rep(1, length(y)), strict = FALSE, OR = 20,
  maxCol = 31, drop.factor.levels = TRUE, nbest = 150,
  user.int = TRUE)
}
\description{
This function returns some necessary information such as residuals
}
\author{
this is a different version of bicreg which was written
by Adrian Raftery, Jennifer Hoeting, Chris Volinsky, Ian Painter, Ka Yee Yeung, 2018
}
